Refactor MapTreeView filenames #225
