+++ visualization/app/codeCharta/ui/codeMap/codeMap.mouseEvent.service.ts
@@ -1,4 +1,4 @@
-import {MapTreeViewHoverEventSubscriber, MapTreeViewLevelController} from "../mapTreeView/mapTreeViewLevelComponent";
+import {MapTreeViewHoverEventSubscriber, MapTreeViewLevelController} from "../mapTreeView/mapTreeView.level.component";
 import {ThreeCameraService} from "./threeViewer/threeCameraService";
 import {ThreeViewerService} from "./threeViewer/threeViewerService";
 import {IAngularEvent, IRootScopeService} from "angular";
